requirements.txt
python-telegram-bot==20.6
gTTS==2.3.2
# fpdf==1.7.2  # Не используется, поэтому закомментировано